Adding the simplification of the noun forms 'realpart, 'imagpart, 'cabs and 'carg.
'realpart and 'imagpart simplifies for real and complex numbers. Expressions which expand to a real or complex number are included.

'cabs and 'carg simplifies for real and complex numbers and expressions for which a positive or negative sign can be determined.

'carg(0) simplifies to 0 to be consistent with the verb form $carg.

The properties verb and noun are put on the property list. This is necessary to get a correct linear display.

Tested with GCL 2.6.8 and CLISP 2.44. No problems with the testsuite and the share_testsuite.