Checking a^2+b^2 not to be zero in the routine sce-int.
The routine sce-int solves integrands of the type: exp(a*x)*sin(b*x) or exp(a*x)*cos(b*x).
When a^2+b^2 is zero, exponentialize the integrand and try the integration again.

Adding some tests for this type of integral to rtest16.mac.

Related bug report: Bug ID: 2797885 - "problem with integration"

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.