Preserve the type of the number zero, when removing a term in PLUSIN.
Do not simplify a 1.0 or 1.0b0 as an exponent away.

One change in the testsuite:
rtest9.mac problem 67: The taylor expansion does not simplify as expected to an expression with floating point numbers. The comparision with an expression in terms of floating point numbers now fails. The result has been adjusted.

No more problems with the testsuite.

Related bug report:
Inconsistent simplification of 1.0*x etc. - ID: 751934