Set limitp not in the routine $defint but in the routine defint.
The answer of $imagpart depends on the value of limitp. In the routine $defint we prepare the integrand and check the boundaries of the integral using the routine $imagpart. To get  the expected answer from $imagpart we do not set limitp in $defint, but later in the routine defint.

Related bug report: defint fails to determine if one of its limit is real - ID: 2840566

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.