Replace shortcuts sin%, cos%, tan%, bess, bessy, kmodbes and gm with simplifying code.
Start to implement the simplification of expressions more systematically. At many places the shortcuts produce unsimplified results, e.g. sin(%pi/2) as part of the result.

Comment out no more used functions and move functions which are no longer used in hypgeo.lisp, but in hyp.lisp to the file hyp.lisp.

No problems with the testsuite and share_testsuite.