Bug report 2876387: ode2 asks if sqrt(3) is an integer
When solving DEs of the form x^2*'diff(y,x,2)+x*'diff(y,x)+(x^2-n^2)*y=0,
we need to know if n is an integer.  Replace existing code in bessel2()
with a call to askinteger function.

This patch does not fix the original problem, which requires inprovements
to askinteger, but it does resolve others found during debugging.