Check the sign with $csign in simpexpt:
$csign checks more carefully if the expression is complex and does not throw a Maxima error like $sign (which is called in noneg). Because $csign only use its complex mode when not in limitp the routines of defint and limit are not affected by this change.

Some examples in rtest_integrate have to be updated, because the sign of the square of a complex symbol is no longer positive, e.g. sign(z^2) -> pz and csign(z^2) - >complex for z a complex symbol. We get more correct noun forms in rtest_integrate.

Related bug report:
Bug ID: 711885 - Rootscontract with imaginaries fails

No problems with the testsuite and share_testsuite.