Reverting the last change to the function compsplt:
The change has been to strong. The functionality of splitprod and splitsum has been switched off completely. These functions are the source of the bug for the sign of (1-(1+sqrt(2))*x).

Making the variables lhs and rhs locally in splitprod and splitsum by renaming them to lhs1 and rhs1:

The real problem is that the variables lhs and rhs are declared to be global. The functions splitprod and splitsum need these variables to be declared locally. The functions call each other and overwrite the temporary result stored in the global variables.

With this changes the following remains correct:

sign(1-(1+sqrt(2))*x) --> pnz
sign(1-(1+sqrt(2))/x) --> pnz

No problems with the testsuite and share_testsuite.

One example in rtest_sign.mac works again.