Reverting in part revision 1.97:
Do not return an unsimplified 0^a expression. The routine timesin cannot handle such an expression. We get a Lisp error for something like 2*0^a. We return ZERO, when the sign of the realpart(a) is not known. That is the behavior of the code before the revision 1.96. Special simplifications are still present, when the sign is known.

No problems with the testsuite and share_testsuite.