Adding code in the routine findthem to check if we have a bounded symbol before doing evaluation. If not return the symbol.
With this change it is possible to use the pattern matcher much more general and it is possible to pass the values of variables to the pattern matcher. E. g. we can write a pattern like `(a free ,var) or can  do a call like (m2 'x (a alike1 x)).

This extension is necessary to write pattern which use arbitrary variables and not only the special VAR.

No problems with the testsuite and share_testsuite.