Removing the use of the global special variable VAR from the code.
The global VAR is used in Schatchen to match pattern. All pattern in this file have been reformulated to depend no longer on the global variable VAR. The variable of the pattern is passed to Schatchen.

The variables *var* and *par* are global to this file only. They are initialized in the routine defexec. The value is never changed. These variables avoid that we have to pass their values through all functions of this file. Furthermore, the use of the variables *var* and *par* is implemented more consistently.

To make the rework of the pattern possible, the functionality of the pattern matcher has been extended with revision 1.14 of schatc.lisp.

No problems with the testsuite and share_testsuite.