Changing the definition of mopl from defmvar to defvar.
This prevents the global list to be reseted by the command reset. mopl holds a list of build-in operators which have got additional properties, but are not allowed to be killed by the function kill-operator.

Related bug report:
ID: 2994324 - tests in rtest_dot fail in 2nd run

No problems with the testsuite and share_testsuite.
The testsuite can be run twice without errors.