Reverting the replacement of exp with expr in the routines superexpt and monstertrig.
We do not get an error with the testsuite or share_testsuite and I have found an example which will fail. But it is more safe not to change the special variable exp to the local variable expr. The reason is, that we call substint in both routines and the routine substint uses the global value of exp.

Correcting a typo. It is method 3 and not method 2 at two places.

No problems with the testsuite and share_testsuite.