1. Replacing the call of the function $listp_check with a call to $listp. 2. Calling merror and not error in the first check.
The function $listp_check is user accessible and not documented. The function is used in mload.lisp to do an assignment check for the variables $file_search and $file_types. But both variables which get the property to check with $listp_check an assignment are not used in Maxima. The function $tcl_output is the only place this function is used outside the file mload.lisp. We cut out $listp_check at this place and replace it with a call to $listp.

No problems with the testsuite and share_testsuite.