With revision 1.123 the function *quo has been replaced with "/" and quot. With the following changes we get the old behavior of the code and no change of the algorithm:
In *red replace "/" with truncate. At this place both arguments are integers.

In addk replace "/" with truncate. At this place both arguments are integers.

In simpquot replace quot with "/". At this place at least one of the arguments is a floating point number.

In exptrl replace quot with truncate. At this place both arguments are integers.

No problems with the testsuite and share_testsuite.