Splitting draw package into two files: * draw.lisp: the object and scene builder * grcommon.lisp: the options handler
Deprecated options:
* 'file_bgcolor': substituted by 'background_color'.
* 'pic_width', 'pic_height', 'eps_width', 'eps_height',
  'pdf_width', and 'pdf_height': all six are substituted
  by new option 'dimensions', also valid for screen
  terminals.
* 'rot_horizontal' and 'rot_vertical': substituted by
  new option 'view'.

When deprecated options are used, the new ones are
automatically called and a warning message is printed.
I am simplifying the draw options system, and this
process will continue in future versions; I'll try
to maintain back portability as much as possible.

Documentation explaining these changes is not yet
written.