Changing the defintion of trisplit from a defmfun-function to a defun-function.
This does not change the functionality of trisplit. The only difference is that the property list of trisplit no longer contains the entry 'translated. This information seems not to be needed in Maxima.

But with this change the command timer(?trisplit) works as expected. The Lisp funtion no longer is killed when doing kill(all). See the related bug report.

Related bug report:
3067098 - The command timer for a Lisp function

No problems with the testsuite and share_testsuite.