Replacing the magic constant -1 with the descriptive variable *parse-stream-eof*
Adding a EOF-test in the routine scan-token. All other routines, which
read a token, do an EOF-test, but not the routine scan-token.  When
reading a file, which ends with a token, this causes an error message 
"The value -1 is not of type CHARACTER.".  The additional test gives
the more expected error message:  "parser: end of file while scanning
expression."