Correcting the function dimnary.
It is an error in dimnary to call the function dimension-paren
directly. This way the global variables are not updated correctly,
because we can not pass the variable w to the function
dimension-paren. When inserting the code of dimension-paren in dimnary
it is possible to update the variables correctly.

This is an old bug. We have two bug reports, which are caused by the
call of dimension-paren from the function dimnary.

Related bug reports:
ID:  635708 - Bad display of unsimplified expr
ID: 1954693 - 2D display of long denominator of nested fractional