Reimplementing the check for a $bfloat in *fpexp
The function is called with an argument, which already has been
converted to a bigfloat. The implementation calls $bfloat a second
time for numbers and for expressions. It is not necessary to do this
for expressions. The second call is only necessary for bigfloat
numbers to adjust the precision of the number. This can be done with a
call to bigfloatp too.