Reimplementing the algorithm for multiplying the power of an integer with a rational or an integer in timesin.
The results have not changed, but the code is more readable, shorter
and uses better the known facts of variables. Therefore, the code
should be more efficient.

Now, the function exponent-of is only called with integer arguments
from the function timesin. The argument m is always an positive
integer and the argument base is guaranteed to be an integer.
Therefore, it is possible to simplify the function exponent-of and to
cut out the tests for other types of arguments. Especially, the call
of the higher level and time consuming function ratgreaterp is no
longer necessary.

No problems with the testsuite and share_testsuite. The testsuite
contains a lot of example which already test the modified part of the
function timesin, because of earlier work on this algorithm.