Increase accuracy of expt for bigfloat args.
When computing x^a via exp(a*log(x)), we need extra bits because the
integer part of a*log(x) doesn't contribute to the accuracy.  So the
computation of exp(a*log(x)) is done with extra precision, and result
is then rounded back to the original precision.  

o Add EXPT-EXTRA-BITS to compute how many extra bits to use.
o Add WITH-EXTRA-PRECISION macro to execute the body with extra
  precision.
o Update EXPT methods to use extra precision.

With these changes, with fpprec=34, 

  (bigfloat:expt (bigfloat:bigfloat 1000) (bigfloat:bigfloat 1000))

returns 1b3000.  Previously maxima returned
+1.000000000000000000000000000000371b3000.