Rewrite the implementation of mat_function and kill matsolve2
This patch replaces the old implementation of mat_function by
something rather more modular and understandable. In case it's of
interest, the result should run faster on large matrices A or for
complicated functions f because it caches the results of the
differentiation and the Taylor expansions for each block.

This also kills the "chlist" and "ghelp" functions, replacing the
first with a histogram function (although it's a simple-minded one,
because we know the list is sorted) and getting rid of the second
entirely.

The new auxiliary functions have names prefixed with "diag_", which
should help avoid name collisions.

We also get rid of "matsolve2", replacing it with
"diag_kernel_element" (which basically does the same thing, but in a
much more understandable way).

The resulting code runs marginally faster and takes marginally less
space (but we're talking 1% or so) on my tests with 4x4 and 5x5
matrices. That's not the point, though. The point is that the
documentation and the slightly more functional style should make it
possible to actually read the implementation now.