Refactor do-merge-asym to remove RES
Instead of having a captured variable, define a macro "WITH-COLLECTOR"
which binds a function to accumulate results. The only users of the
old RES were via (lambda (x) (push x res)), so this will run exactly
the same speed.

This gets rid of a warning when compiling compar.lisp because we were
calling do-merge-asym purely for the side effects and thus discarding
the result of an nreverse.