Significantly refactor ASKSIGN1
The result pretty much disembowels ASKSIGN1, putting its contents in
two new functions: ENSURE-SIGN and MATCH-SIGN. MATCH-SIGN has the
parsing logic ("you say this and I'm expecting that, so the sign
is..") and ENSURE-SIGN calls it repeatedly, supplying data from a call
to RETRIEVE if it needs to ask the user.

As such, we can remove the slightly-modified-copy of ASKSIGN1 from
share/contrib/noninteractive/asksign1.lisp. Instead, asksign1.lisp
slightly modifies ENSURE-SIGN to throw if a single call to MATCH-SIGN
fails.

We also rename the special variable LOCALS in compar.lisp
to *LOCAL-SIGNS* (since that's what it keeps track of). (Finding users
of the LOCALS variable for the rename was what prodded me to make the
other change)

The result runs the testsuite successfully. It also runs the
rtest_noninteractive testsuite as successfully as it did before the
change.