Disable portability checks in automake
The build doesn't work without GNU Make anyway, so we may as well not
get spammed with complaints about it. This patch also moves to the
"new" autoconf syntax. A quick hunt online didn't find when this
syntax was first added, but a message from 2003 [1] described it as
the preferred syntax then, so I'm pretty certain I'm not introducing a
dependency problem!

[1] https://lists.gnu.org/archive/html/bug-automake/2003-04/msg00000.html