Add a configure flag --enable-sys-proclaim
The default is false and the user is warned if they enable it without
GCL (since that wouldn't actually do anything).

Note that this patch deletes some conditional WIN32 code. As far as I
can see, the only difference between the two versions was the order in
which we enabled emit-fn and loaded defsystem. Since we no longer want
to enable emit-fn at all, there is no longer any difference between
the two.

Not yet done with this patch: Force regeneration of sys-proclaim.lisp
when a lisp file changes.