Gather together limit-related tests
rtest16 was enormous, so this gives a nice partial
categorisation. This sort of thing seems useful since, when I break
something in the testsuite, it's convenient to just be able to
repeatedly run one file full of failing tests. Now if someone breaks
limit(), they'll get a screenful from that file...