Make ModeMatrix more convenient to call.
If the jordan_info argument is not given, it's calculated for the
user. The ridiculous nested if's in ModeMatrix are kind of unpleasant,
but I think we need them for noun forms + optional arguments to work
correctly. I don't think there's a version of apply('foo, args) that
does what I want.