Remove the [$]SUPERLOGCON variables
These were declared as special, but don't seem to do anything. They
only occurred once within the tree other than in the DEFMVAR forms and
a special declaration in lmdcls.lisp. This occurrence was in LOGCON,
where setting $SUPERLOGCON to false would have disabled the LGCSORT
phase of the log-contract algorithm.

Since SUPERLOGCON wasn't documented anywhere and it seems that it's
only effect was to cripple an otherwise working function, this patch
removes it.