Fix bug in trigreduce with complicated expressions
This fixes bug 2594. In particular, if you define

  testit(a, n) := block([initial: product(cos(k*x), k, 1, n)],
                    [subst(a, x, initial),
                     subst(a, x, trigreduce(initial))])$

then testit(0,k) gave the right answer for k=1,...,7, but gave
something silly for k=8.

It turns out that there was a logic error in sp1tplus. That function
calls sp1sintcos repeatedly and gets back either a term of the form
sin(foo) or cos(foo) or it gets a sum of such terms. In the latter
case, if it started with a coefficient on the front, it might end up
multiplying the coefficient by the sum. Unfortunately, this meant that
on the next time round, the term wasn't in the right form and
sp1sintcos started giving silly answers. The new code is much more
paranoid and explicitly throws an error if it gets a result in a form
that it doesn't recognise.