Fix missing DESTDIR in Makefile.am's
Pointed out by Rex Dieter, I accidentally broke the DESTDIR support
for clisp and scl. This fixes that mistake, and also adds destdir to
the extradocinstall target in the top-level Makefile.am. Probably that
only ever gets run on Windows, so DESTDIR is never set, but we may as
well be consistent.