Get rid of errrjfflag
This worked by setting a special variable (errrjfflag) to true and
then, when there was a rational function error, some code threw to a
hopefully existing 'raterr tag, rather than calling merror.

This patch replaces this mechanism by using the CL condition
system. We define a condition class called rat-error and places that
used to throw to raterr now raise a rat-error condition. Instead of
binding errrjfflag to true and adding a catch block, calling code now
just handles the raised rat-error (made easy by the ignore-rat-err
macro).

Finally, the top-level repl catches rat-errors in the same place as it
with the code that catches other random lisp errors. However it
special cases these ones and calls merror with the given
arguments. The idea is that a more general Maxima condition system
will probably look something like this and it will be easy to port
this code to however that might turn out.

Incidentally, this is already slightly badly named. For example,
solcoef complains about something being nonlinear via this
mechanism. But we may as well keep the terminology for now...