Document ALG, PACOEFP and LEADTERM and get rid of CREMAINDER.
The latter was only used in two places, and was used for two rather
different things. In particular, it didn't actually return a remainder
as such: when MODULUS was true, (cremainder a b) => 0, even though for
non-prime modulus this doesn't really make sense... Apparently, this
was set up for its use in the pseudo-remainder function PREM, so I
moved the logic to there.