Improvements to CRECIP
When trying to understand and document CRECIP, I wasn't entirely
convinced about the fixnum-specific version, so I had a look at the
generated code in SBCL and realised that it only half succeeds in
getting non-generic arithmetic. This is mainly (I think) because of
the product that it has to compute: if you merely tell lisp that the
factors are fixnums, it doesn't know that the product has to be.

The new code is (I think) a lot easier to read. It has fixnum-specific
code in its function body and in the case of large modulus, it calls
out to a generic version. Sadly, it's slightly larger: the code for
the new CRECIP weighs in at 1090 bytes, plus 474 for CRECIP-GENERAL,
compared to the old one that came to 1224. Most of the bloat seems to
be SBCL putting type checks at the start and error handlers at the
end.

On the plus side, it's about 25% faster. Calling the old version
CRECIP-OLD, I ran the following test:

MAXIMA> (let ((pairs
               (delete nil
                       (loop
                          for i below 10000
                          collect (let ((mod (+ (random 5000) 2))
                                        (sgn (- (* (random 2) 2) 1)))
                                    (when (primep mod)
                                      (cons mod (* sgn (1+ (random (1- mod)))))))))))
          (print (length pairs))
          (time
           (dotimes (k 10000)
             (mapc (lambda (pair)
                     (let ((modulus (car pair)))
                       (crecip (cdr pair)))
                     nil)
                   pairs)))
          (time
           (dotimes (k 10000)
             (mapc (lambda (pair)
                     (let ((modulus (car pair)))
                       (crecip-general (cdr pair)))
                     nil)
                   pairs)))
          (time
           (dotimes (k 10000)
             (mapc (lambda (pair)
                     (let ((modulus (car pair)))
                       (crecip-old (cdr pair)))
                     nil)
                   pairs))))

1327
Evaluation took:
  1.890 seconds of real time
  1.908000 seconds of total run time (1.908000 user, 0.000000 system)
  100.95% CPU
  4,336,418,928 processor cycles
  16,112 bytes consed

Evaluation took:
  3.353 seconds of real time
  3.384000 seconds of total run time (3.384000 user, 0.000000 system)
  100.92% CPU
  7,693,280,708 processor cycles
  18,544 bytes consed

Evaluation took:
  2.524 seconds of real time
  2.544000 seconds of total run time (2.544000 user, 0.000000 system)
  100.79% CPU
  5,793,591,732 processor cycles
  16,360 bytes consed

Looking at the generated code, we could probably do even better if
SBCL knew that MERROR never returned, for example. I'll probably look
at that type inference next.