Remove #-kcl guard
This still applied with GCL, but removing it appears to have no
effect. Calling (describe 'cmod) or (describe 'maxima::cmod) results
in a description of a symbol defined in the system package. Run times
on testsuite are obviously not massively affected:

    |            |    Time | GC time |
    |------------+---------+---------|
    | #-kcl      | 217.460 |  30.659 |
    | No #-kcl   | 214.410 |  27.559 |

Since they go in the opposite direction from what one might fear, I
assume this means that the results are lost in the noise. (No, I
wasn't using the system for anything else at the time) I have no
intention of watching GCL for an hour in the hope that the numbers
will stabilise, so assume that this just works.