Get rid of the MCMOD macro and document CMOD
The only users were CPLUS, CTIMES and CDIFFERENCE and they could be
refactored to just call CMOD. The cost of this is a function call that
was open-coded when MCMOD was a macro. However, this will be dwarfed
by the cost of the generic arithmetic functions that they are doing,
so duplicating the code seems silly.

I can't detect an effect of the change on run times for the test
suite, either on SBCL or on GCL. Run times on SBCL seem to have
dropped slightly, but this could just be natural variation.