Rename SETQMODULUS to SET-MODULUS and document
The rename is because I spent a while pondering what "qmodulus" was
before I realised it's "setq modulus". The other change here is to
replace the MERROR call with a hard lisp error: looking at the code,
this isn't called when a user binds modulus at the repl, so a bogus
value here means a coding error in lisp code. As such, it should throw
a proper lisp error, I think.