Remove bizarre comment
This isn't really quite true. The functions that are called
"CSOMETHING" do indeed expect the coefficient to be a
number (something for which pcoefp returns true). The logand note is
true, but only applies to a single occurrence where both my new code
and the old version are careful to check that it will work.