Get rid of errrjfflag (take 2)
This worked by setting a special variable (errrjfflag) to true and
then, when there was a rational function error, some code threw to a
hopefully existing 'RATERR tag, rather than calling MERROR.

The previous version of this patch replaced that logic by using the CL
condition system. Unfortunately, this turned out to be ridiculously slow
on CMUCL. The new version uses CATCH and THROW again, but hides the
implementation behind some tidy looking macros.

I've just run the testsuite with and without this patch and got the
following timings on CMUCL (in seconds)

  Without patch:  213 214 216
  With patch:     212 215 217