Re-add src/share-subdirs.lisp
This was deleted in commit 320b15e. I guess that was probably a mistake,
but it's arguably the right thing to do (since we don't really want to
support non-autotools builds any more?). I'm going with this option
since it's probably less controversial, but I'm also happy with adding a
line to gitignore :-)