Flatten out the not-quite-tail-recursion in PEXPT2
The reason I was looking at this was to try to document it, and I spent
ages figuring out exactly what "PEXPT2" did, carefully calculating how
the recursion worked. Once I'd finally understood it enough to write
documentation, I realised that it was would be easier to understand if I
rewrote it with a DO loop. And it shouldn't hammer the stack quite as
much. Testsuite times don't change appreciably.

The other nice thing is this removes *a* and *x*, the last of the
special variables declared in rat3a.lisp. Woo!