Localise *prompt-prefix* and *prompt-suffix* use
This has all been rationalised into a function called FORMAT-PROMPT. The
idea is that the next commit will add further ways in which a user can
customise what appears on the prompt and I want to make sure that it
works correctly.

This commit also pretty much rewrites AFORMAT to just punt to MFORMAT,
possibly wrapping the output in a string. As far as I can tell, that's
basically what it did already, except it worked by including the guts of
the function via the MFORMAT-LOOP macro. And it errored with destination
equal to 'T', which seemed less than helpful...