Document PSIMP
There are basically no changes to the code. PSIMP1 was only used in
PSIMP2, and could be elided. PSIMP2 is a destructive operation that
deletes terms with zero coefficients from a polynomial. The code isn't
changed at all, but some of the c???r calls get replaced with the
equivalent pt-lc and pt-red macros to make it a bit clearer what's going
on.