Fix a performance regression in the last commit.
Fortunately, I checked for this before I pushed it. Before commit
1268326, a testsuite run took 181.2 - 181.4 seconds on my machine with
SBCL 1.1.13. After it, the times jumped up to 182.4 - 182.6 or so.

I was mightily confused, because I thought that the code was pretty much
equivalent. The confusion got stronger when I started poking around: the
old and new versions of PSIMP compile to identical code. It turns out
that the change was in PSIMP2 (which became P-DELETE-ZEROS). In order to
elide PSIMP1, I wrapped the whole of PSIMP2 in a (let (($algebraic))
...)  form. It seems binding this special is quite expensive and isn't
needed in most calls to P-DELETE-ZEROS.

With this patch, the binding gets moved down to the recursive PSIMP call
where it's actually used. Test suite times are now between 179.8 and
180.2. Very satisfying!