On windows platforms, add path to gnuplot if required.
This is a bit of a hack, but mgnuplot isn't that pretty anyway.
On windows, check to see if gnuplot is in the path.  If not,
then assume that we are runing as part of the windows installer package
and set the path relative to the mgnuplot.exe executable.

Setting PATH here is the least worst solution I could find.