Complete migration of plot options into a Lisp property list
This completes the migration from the old global variable $plot_options to a global variable *plot-options*, which had been started several years ago. This new method makes it easier to maintain and add new plot options. Maxima functions get_plot_option and set_plot_option continue to work and get_plot_option can now be called without any arguments, giving the list of all active options.

From a lisp program, it is more convenient to use (getf *plot-options* :option-name) to get the value of an option, (setf (getf *plot-options* :name) new-value) to set an option value and (remf *plot-options* :name) to remove options. All options passed to a plotting program can also be given to program plot-options-parser which will check that they are known and valid options and add them to the list of options.

The major difference with the old method is that users can no longer change the options directly by changing the value of plot_options, but they have to use set_plot_option. This will prevent the creation of malformed plot_options variables. The global variable plot_options has been kept, with a minimal content, to avoid failure of external packages that might use it, but none of the plotting packages in Maxima are using it.

This is a large commit, but the use of $plot_options was so extended that it was no possible to do it in smaller commits. I have tested with all the figures in the plotting chapter of the manual, using 3 different Lisp flavors.