Fixes a bug with the gnuplot_term option
In older versions, users got use to giving values of default or "png" to option gnuplot_term. If they use "default" or png, that would not work. This has become evident in this new version where the type of arguments that each option must have is been checked.

gnuplot_term is a tricky option to handle, because when it is just default, dumb or ps, plot.lisp expects it to be a symbol, but when it is more complicated, for instance, "ps; size 16cm, 12cm", it must be a string and not a symbol. Rather than changing the many places where gnuplot_term appears in plot.lisp, the plot-options-parser function will make sure that when it is one of the 3 key terminals (default, ps or dumb) it will cahnged to a symbol. In other cases, the value entered will be forced to be a string and the user will be responsible to make sure that the result makes sense to Gnuplot.