Recover the exact same behavior of get_plot_option as in older versions
There is no need to quote strings in the output. I was confused thinking that results such as [gnuplot_dumb_term_command, set term dumb 79 22] could not be used later on to set the same option, but it is not true; set_plot_options(get_plot_options()) works fine without quoting. On the other hand, the extra quotes led to errors in Wmaxima.

get_plot_option gives results s