Fixes texinfo macro definitions
The previous definition of the macros led to an error when using Texinfo 5, making "make" fail.

According to section 19.3 (Macro Details and Caveats) of the Texinfo manual:
"Macros containing a command which must be on a line by itself, such as a conditional, cannot be invoked in the middle of a line."
The commands @ifhtml and @ifnothtml have to be on a line by themselves so they could not be inside the mref macros which are invoked in the middle of lines.

I have checked this new version of category-macros.texi both in Texinfo 4 and 5, and it worked in both.