Deal more robustly with floating point errors
When we would have got a floating point exception, build the relevant
bigfloat instead.

  (%i180) 1000! * 0.01;

  (%o180) 4.023872600770938b2565
  (%i181) 1000! + 0.01;

  (%o181) 4.023872600770938b2567

This works on every supported lisp except Allegro, where handling a
type-error doesn't work. I intend to push this (which doesn't make
things worse on Allegro). If the missing functionality is annoying an
Allegro user, maybe someone with more experience of Franz lisps can
figure out a patch.

This fixes bug 2680 on non-Allegro lisps and, I suspect, helps lots of
other confused users. Searching the web for the strings that SBCL gives
on floating point overflow returns hits that include up Sage and
Launchpad bug reports about Maxima!

My logic for why this is "clearly the right thing to do" is that Lisp
has arbitrary sized integers. Maxima also has arbitrary sized floats, so
we should try to use them to avoid things exploding. I was a little
worried that there would be a massive performance hit, but was
pleasantly surprised.

Test suite timings slow down slightly on GCL (4%), CMUCL (3%). On the
other implementations, there doesn't seem to be a dramatic
difference. Bizarrely, CCL seems to get 2% faster! On those lisps that
report it, the number of bytes consed usually goes down with the patch,
by 0.5% or so.

Note that the "after patch" row doesn't include tests 251-253 of
rtest15.mac which I removed (because eg. 2*foo is no longer NaN), but I
hardly think they take 13 seconds...