Trivial change to fix some subscripted integration
$DEFINT used STRIPDOLLAR to make a temporary integration variable when
you passed it a subscripted variable. This turns out to be less than
brilliant... The first example in #2676 hit this bug because it was
using x[1]. It turns out that X is the dummy variable used in the
integral lookup table (and (STRIPDOLLAR $X) => X).

Gensyms seem a saner way to make symbols...