Remove code duplication of the hypergeometric representation of the Bessel functions.
The hypergeometric representation of the Bessel functions bessel_j,
bessel_y, bessel_i and bessel_k is used when the order is complex,
and it's also returned when hypergeometric_representation is true.
This commit removes the duplication that was present between these
two cases for each function.

The test suite runs fine.