Fix build of pt_BR docs
Since I don't speak a word of Portuguese, let alone a Brazilian flavour,
I'm a little wary about editing these... But all I've done is moved two
subsubsections up to subsection level. These were directly contained in
a section, which wouldn't compile with the version of texinfo I have
here.

The German docs don't compile on my system at the moment. I do speak a
little German, but the torrent of error messages on the screen in front
of me suggests that fixing that might be rather more work!