Add the promote_float_to_bigfloat option variable
This controls automatic promotion on floating point overflow. The patch
contains code that (I think!) makes everything work, together with a
simple test and some documentation.

While writing the documentation, I noticed the float2bf variable (that
I'd never noticed before). It occurred to me that maybe we should just
use that as a flag instead. "float2bf" is a generic enough name that it
describes this operation too, and maybe this is a good way to avoid yet
another flag. On the flip side, flags with multiple magic properties are
confusing...

I'll push this for now, and copy the above to the mailing list for
discussion. Hopefully we can decide the best way to structure this
before the next release.