%e^^A should not simplify to %e^A
When A is a literal matrix, the simplification happened because of some
incorrect logic in SIMPNCEXPT. I think this logic is supposed to allow
simplifications like

  u^^3 = u^3

when u is known to be a scalar. I guess that if u=v*w, with v and w
scalar, this is be useful. Unfortunately, it doesn't deal with the
possibility that ^^ is being used to raise by a matrix.

With this patch, the matrix example gives a noun form (which, although
not particularly helpful, does have the advantage of not being wrong!)